+++ resources/js/Pages/Collections/CollectionOfTheMonth.tsx
@@ -3,7 +3,7 @@ import { Head } from "@inertiajs/react";
 import cn from "classnames";
 import { useTranslation } from "react-i18next";
 import { IconButton } from "@/Components/Buttons";
-import { WinnersChart } from "@/Components/Collections/CollectionOfTheMonthWinners";
+import { WinnersChart } from "@/Components/Collections/CollectionOfTheMonthWinners/CollectionOfTheMonthWinners.blocks";
 import { Heading } from "@/Components/Heading";
 import { Icon } from "@/Components/Icon";
 import { Link } from "@/Components/Link";
